IVYPORTAL-11673 Various or urgent requests in IVYPORTAL S203 - Mark deprecated method for Portal 8, 9

+++ AxonIvyPortal/PortalKit/src/ch/ivy/addon/portalkit/util/DateTimeFormatterUtils.java
@@ -22,7 +22,7 @@ public class DateTimeFormatterUtils {
   }
 
   private static String prepareShortYearDisplayText(long days) {
-    long years = (long) days / DAYS_IN_A_YEAR;
+    long years = days / DAYS_IN_A_YEAR;
     long remainingDays = days - years * DAYS_IN_A_YEAR;
     String dayDisplayText = remainingDays > 0
         ? DASH_SPACE + prepareShortDateTimeDisplayText(remainingDays, CmsDateTimeUnitService.getCmsDayUnit())
